refactor(dashboard): extract greeting/header/suggestion/season into useDashboardCopy
Move greetName, headerHook, suggestion, and the Dec/Jan wrappedSeason
state+effect verbatim into _hooks/use-dashboard-copy.ts. Owns its own
useTranslations("Dashboard.body"); takes the derived data it needs.
Behavior-identical; page.tsx 1625->1558 LOC.
